Use bigint for channel ids

+++ database/migrations/0003_create_subscriptions.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS channel_subscription (
-  channel_id INTEGER NOT NULL,
+  channel_id BIGINT NOT NULL,
   expediente_id VARCHAR(255) NOT NULL,
   UNIQUE (channel_id, expediente_id)
 );
